Wave-1 integration: wire platform into tests/ foreach; clang-tidy cleanup across merged a-theta modules (const-correctness/uppercase-suffix/isolate-decl/narrowing-cast, 124 findings); WORKFLOW Wave-1 retro + W1-09/10/11 backlog + C1 handoff

+++ src/geom/geom.cc
@@ -267,7 +267,7 @@ void geom_out_init(struct GeomOut* o, struct TVtx* tverts, uint32_t tvert_cap,
   o->tris_total = 0;
   o->tris_dropped = 0;
   for (int i = 0; i < GEOM_NUM_TILES; ++i) {
-    o->tiles[i].refs = &refs[(uint32_t)i * refs_per_tile];
+    o->tiles[i].refs = &refs[(size_t)i * refs_per_tile];
     o->tiles[i].count = 0;
     o->tiles[i].cap = refs_per_tile;
     o->tiles[i].dropped = 0;
